+++ release/splitTools/R/partition.R
@@ -1,12 +1,12 @@
 #' Split Data into Partitions
 #'
-#' This function provides row indices for data splitting, e.g. into training, validation, and test. Different types of split strategies are supported ("basic", "stratified" (the default), or "grouped"), see Details. The partition indices are either returned as a list with one element per partition (the default) or as vector of partition ids.
+#' This function provides row indices for data splitting, e.g. into training, validation, and test. Different types of split strategies are supported ("basic", "stratified" (the default), "grouped", or "blocked"), see Details. The partition indices are either returned as a list with one element per partition (the default) or as vector of partition ids.
 #'
-#' By default, the function uses stratified splitting. This will balance the partitions as good as possible regarding the distribution of the input vector \code{y}. Numeric input is first binned into \code{n_bins} quantile groups. If \code{type = "grouped"}, groups specified by \code{y} are kept together when splitting. This is relevant for clustered or panel data.
+#' By default, the function uses stratified splitting. This will balance the partitions as good as possible regarding the distribution of the input vector \code{y}. Numeric input is first binned into \code{n_bins} quantile groups. If \code{type = "grouped"}, groups specified by \code{y} are kept together when splitting. This is relevant for clustered or panel data. In contrast to basic splitting, \code{type = "blocked"} does not sample indices at random, but rather keeps them in groups: e.g. the first 80% of observations form a training set and the remaining 20% are used for testing.
 #' @importFrom stats ave quantile
-#' @param y Either the variable used for "stratification" or "grouped" splits. For "basic" splits, any vector of the same length as the data intended to split.
+#' @param y Either the variable used for "stratification" or "grouped" splits. For other types of splits, any vector of the same length as the data intended to split.
 #' @param p A vector with split probabilities per partition, e.g. c(train = 0.7, valid = 0.3). Names are passed to the output.
-#' @param type Split type. One of "stratified", "basic", "grouped". The default is "stratified".
+#' @param type Split type. One of "stratified", "basic", "grouped", "blocked". The default is "stratified".
 #' @param n_bins Approximate numbers of bins for numeric \code{y} and \code{type = "stratified"}.
 #' @param split_into_list Should the resulting partition vector be split into a list or not? Default is\code{TRUE}.
 #' @param use_names Should names of \code{p} be used as partition names? Default is \code{TRUE}.
@@ -22,8 +22,10 @@
 #' partition(y, p, split_into_list = FALSE, seed = 1)
 #' partition(y, p, split_into_list = FALSE, use_names = FALSE, seed = 1)
 #' partition(y, p = c(0.7, 0.3), type = "grouped")
-partition <- function(y, p, type = c("stratified", "basic", "grouped"), n_bins = 10,
-                      split_into_list = TRUE, use_names = TRUE, seed = NULL) {
+#' partition(y, p = c(0.7, 0.3), type = "blocked")
+partition <- function(y, p,
+                      type = c("stratified", "basic", "grouped", "blocked"),
+                      n_bins = 10, split_into_list = TRUE, use_names = TRUE, seed = NULL) {
   # Input checks
   type <- match.arg(type)
   stopifnot(length(p) >= 1L, p > 0,
@@ -39,6 +41,8 @@ partition <- function(y, p, type = c("stratified", "basic", "grouped"), n_bins =
   # Calculation of partition ids
   if (type == "basic") {
     out <- .smp_fun(n, p)
+  } else if (type == "blocked") {
+    out <- rep.int(seq_along(p), times = ceiling(p * n))[seq_len(n)]
   } else if (type == "stratified") {
     if (is.numeric(y) && length(unique(y)) > n_bins) {
       y <- .bin(y, n_bins)