+++ NEWS.md
@@ -1,6 +1,10 @@
 # splitTools 0.2.0
 
-- Added `type = "blocked"` to allow for blocked splitting.
+## New Functionality
+
+- Added `type = "blocked"` to `create_folds` and `partition` to allow for blocked splitting.
+
+- Added function `create_timefolds` for cross-validation of time series data.
 
 # splitTools 0.1.0